moving files parser fns to read_files.py
